refactor: Remove useless CN/OS overrides
