--- src/main/java/ovh/akio/hmu/games/os/HonkaiImpactGameOS.java
@@ -1,20 +1,0 @@
-package ovh.akio.hmu.games.os;
-
-import ovh.akio.hmu.games.cn.HonkaiImpactGameCN;
-
-import java.io.File;
-
-public class HonkaiImpactGameOS extends HonkaiImpactGameCN {
-
-    public HonkaiImpactGameOS(File basePath) {
-
-        super(basePath);
-    }
-
-    @Override
-    public String getName() {
-
-        return "Honkai Impact 3rd (OS)";
-    }
-
-}
