+++ src/main/java/ovh/akio/hmu/interfaces/HoyoverseGame.java
@@ -4,11 +4,9 @@ import ovh.akio.hmu.entities.PckAudioFile;
 import ovh.akio.hmu.enums.Game;
 import ovh.akio.hmu.exceptions.InvalidGameDirectoryException;
 import ovh.akio.hmu.games.cn.GenshinImpactGameCN;
-import ovh.akio.hmu.games.cn.HonkaiImpactGameCN;
-import ovh.akio.hmu.games.cn.HonkaiStarRailCN;
+import ovh.akio.hmu.games.HonkaiImpactGame;
+import ovh.akio.hmu.games.HonkaiStarRail;
 import ovh.akio.hmu.games.os.GenshinImpactGameOS;
-import ovh.akio.hmu.games.os.HonkaiImpactGameOS;
-import ovh.akio.hmu.games.os.HonkaiStarRailOS;
 
 import java.io.File;
 import java.util.List;
@@ -21,34 +19,21 @@ public interface HoyoverseGame {
         try {
             return new GenshinImpactGameOS(path);
         } catch (InvalidGameDirectoryException ignore) {
-
-        }
-
-        try {
-            return new HonkaiStarRailOS(path);
-        } catch (InvalidGameDirectoryException ignore) {
-        }
-
-
-        try {
-            return new HonkaiImpactGameOS(path);
-        } catch (InvalidGameDirectoryException ignore) {
         }
 
         try {
             return new GenshinImpactGameCN(path);
         } catch (InvalidGameDirectoryException ignore) {
-
         }
 
         try {
-            return new HonkaiStarRailCN(path);
+            return new HonkaiStarRail(path);
         } catch (InvalidGameDirectoryException ignore) {
         }
 
 
         try {
-            return new HonkaiImpactGameCN(path);
+            return new HonkaiImpactGame(path);
         } catch (InvalidGameDirectoryException ignore) {
         }
